Extract PurchaseUpgradeErrorView
