+++ WooCommerce/Classes/ViewModels/InAppPurchases/WooWPComPlan.swift
@@ -1,6 +1,6 @@
 import Yosemite
 
-struct WooWPComPlan {
+public struct WooWPComPlan {
     let wpComPlan: WPComPlanProduct
     let wooPlan: WooPlan
     let hardcodedPlanDataIsValid: Bool
